Fixed tests for startup listener
